requirements.txt
pyspark
kafka-python
transformers
elasticsearch==8.13.2
faker
ipykernel
pandas
pyyaml
yfinance
requests
pyyaml